Added Update User , Refactor Base class

+++ src/main/java/com/kampus/service/concretes/UserService.java
@@ -1,37 +1,40 @@
 package com.kampus.service.concretes;
 
+import com.kampus.core.config.JwtService;
 import com.kampus.core.service.BaseService;
+import com.kampus.core.utilities.exceptions.BusinessException;
 import com.kampus.core.utilities.mappers.ModelMapperService;
 import com.kampus.dataAccess.UserRepository;
 import com.kampus.entities.User;
-import com.kampus.service.requests.userRequests.CreateUserRequest;
 import com.kampus.service.requests.userRequests.UpdateUserRequest;
+import com.kampus.service.responses.authentication.AuthenticationResponse;
 import com.kampus.service.responses.userResponses.GetAllUsersResponse;
 import com.kampus.service.responses.userResponses.GetUserByIdResponse;
 import com.kampus.service.rules.UserBusinessRules;
+import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.function.Consumer;
 
+import static com.kampus.core.handler.BusinessErrorCodes.USER_NOT_FOUND;
 
 @Service
 @AllArgsConstructor
-public class UserService implements BaseService<CreateUserRequest, Long, GetUserByIdResponse, GetAllUsersResponse, UpdateUserRequest> {
+public class UserService extends BaseService<Void, Long, GetUserByIdResponse, GetAllUsersResponse, UpdateUserRequest> {
     private final UserRepository userRepository;
     private final ModelMapperService modelMapperService;
     private final UserBusinessRules userBusinessRules;
-
-    @Override
-    public void add(CreateUserRequest createUserRequest) {
-        User user = this.modelMapperService.forRequest().map(createUserRequest, User.class);
-
-        this.userRepository.save(user);
-    }
+    private final JwtService jwtService;
 
     @Override
     public GetUserByIdResponse getById(Long id) {
-        User user = this.userRepository.findById(id).orElse(null);
+        User user = this.userRepository.findById(id).orElseThrow(() -> new BusinessException(
+                USER_NOT_FOUND.getCode(),
+                USER_NOT_FOUND.getHttpStatus(),
+                USER_NOT_FOUND.getDescription()
+        ));
 
         return this.modelMapperService.forResponse()
                 .map(user, GetUserByIdResponse.class);
@@ -42,17 +45,62 @@ public class UserService implements BaseService<CreateUserRequest, Long, GetUser
         List<User> users = this.userRepository.findAll();
 
         return users.stream()
-                .map(user -> this.modelMapperService.forResponse().map(user, GetAllUsersResponse.class)).toList();
+                .map(user -> this.modelMapperService.forResponse().map(user, GetAllUsersResponse.class))
+                .toList();
     }
 
+    // Kullanıcıyı güncellenmiş bilgilerle güncelle
     @Override
-    public void update(UpdateUserRequest updateRequest) {
-        User user = this.modelMapperService.forRequest().map(updateRequest, User.class);
+    @Transactional
+    public AuthenticationResponse update(UpdateUserRequest updateRequest) {
+        // Kullanıcıyı mevcut username ile bul
+        User user = this.userRepository.findByUsername(updateRequest.getCurrentUsername())
+                .orElseThrow(() -> new BusinessException(
+                        USER_NOT_FOUND.getCode(),
+                        USER_NOT_FOUND.getHttpStatus(),
+                        USER_NOT_FOUND.getDescription()
+                ));
+
+        // İş kurallarını kontrol et
+        userBusinessRules.checkIfUsernameExists(updateRequest.getUsername());
+        userBusinessRules.checkIfMailExists(updateRequest.getEmail());
+        userBusinessRules.checkIfPhoneNumberExists(updateRequest.getPhoneNumber());
+
+        // Güncellenmiş alanları ayarla
+        updateUserField(user::setUsername, updateRequest.getUsername());
+        updateUserField(user::setPassword, updateRequest.getPassword());
+        updateUserField(user::setEmail, updateRequest.getEmail());
+        updateUserField(user::setFirstName, updateRequest.getFirstName());
+        updateUserField(user::setLastName, updateRequest.getLastName());
+        updateUserField(user::setPhoneNumber, updateRequest.getPhoneNumber());
+
+        // Güncellenmiş kullanıcıyı veritabanına kaydet
         this.userRepository.save(user);
+
+        // JWT token oluştur
+        String jwtToken = jwtService.generateToken(user);
+
+        // Yanıtı döndür
+        return AuthenticationResponse.builder().token(jwtToken).build();
     }
+
 
     @Override
     public void delete(Long id) {
+        if (!this.userRepository.existsById(id)) {
+            throw new BusinessException(
+                    USER_NOT_FOUND.getCode(),
+                    USER_NOT_FOUND.getHttpStatus(),
+                    USER_NOT_FOUND.getDescription()
+            );
+        }
         this.userRepository.deleteById(id);
+    }
+
+
+    private void updateUserField(Consumer<String> setter, String value) {
+        if (value != null && !value.isEmpty()) {
+            setter.accept(value);
+        }
     }
 }
